Restructure plugin, add stylua
Making it easier to start adding some planned features, and generally
attempting to follow best-practices from this source:

https://github.com/nvim-neorocks/nvim-best-practices
